refactor: restructure database access layer
