+++ cmd/radar/wrangler.toml
@@ -16,17 +16,31 @@ port = 4242
 [observability]
 enabled = true
 
+[triggers]
+crons = ["0 */1 * * *"]
+
+
 [[d1_databases]]
-binding = "DB" # available in your Worker on env.DB
-database_name = "motr-controller-db"
-database_id = "872a4b08-7e07-4978-b227-5b60940238ed"
+binding = "ACTIVITY_DB"
+database_name = "motr-activity"
+database_id = "a7ccb4bb-c529-4f42-8029-92564a3aecb8"
+
+[[d1_databases]]
+binding = "NETWORK_DB"
+database_name = "motr-network"
+database_id = "acb75499-3502-4052-9604-263a913e077a"
+
+[[d1_databases]]
+binding = "USERS_DB"
+database_name = "motr-users"
+database_id = "8ed4d399-5932-419c-b92f-9c20d7a36ad2"
 
 [[kv_namespaces]]
-binding = "SESSIONS" # available in your Worker on env.KV
+binding = "SESSIONS_KV"
 id = "ea5de66fcfc14b5eba170395e29432ee"
 
 [[kv_namespaces]]
-binding = "HANDLES" # available in your Worker on env.KV
+binding = "HANDLES_KV"
 id = "271d47087a8842b2aac5ee79cf7bb203"
 
 [[r2_buckets]]
