+++ internal/db/activity/query.sql
