+++ internal/db/activity/schema.sql
