+++ internal/db/network/models.go
