+++ internal/db/network/query.sql
