+++ internal/db/sqlc.yaml
